api/requirements.txt
python-dotenv
fastapi
sqlalchemy
uvicorn 
psycopg2-binary